pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitbro"
version = "2.1.0"
description = "AI-Powered Git CLI Tool for enhanced developer workflow"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["git", "ai", "cli", "commit", "openai", "ollama", "gemini", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "click>=8.0.0",
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "google-generativeai",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.scripts]
gitbro = "src.cli:main"

[project.urls]
Homepage = "https://github.com/jemmy9211/gitbro"
Repository = "https://github.com/jemmy9211/gitbro"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

